// expense-mate-ui/src/app/core/models/income.model.ts
export interface Income {
  incomeId?: number;
  IncomeId?: number;
  id?: number;
  incomeCategory?: string;
  IncomeCategory?: string;
  category?: string;
  Category?: string;
  incomeAmount?: number;
  IncomeAmount?: number;
  amount?: number;
  Amount?: number;
  createdAt?: string | Date;
  CreatedAt?: string | Date;
  incomeDate?: string | Date;
  IncomeDate?: string | Date;
}